Rewrite tests using pytest
