+++ .travis.yml
@@ -7,4 +7,7 @@ python:
   - "3.3"
   - "3.4"
 
+install:
+  - pip install pytest
+
 script: ./setconf.py --test
